requirements.txt
kafka-python
python-dotenv
valkey
zmq
lnhistoryclient==3.0.1